Add tooltips to panel components. Tweak layout.
To prevent ugly v-if v-else statements and duplicating everything, all
panel components were rewritten to not use templates and are now
javascript components.

Changed the height of the toolbar in small mode to 40px to better fit
the toolbar-options component. Made the render mode icons only show up
when render mode is selected.