Work on editor plumbing.
Instead of emitting editor-open, just set editorItem in the store to
open a file in the editor.

Change signature of editorSave on the global plugin to just return a
promise instead of having to provide resolve / reject callbacks.

Fix destroying the current editor component by replacing it with an
empty <div/> instead of setting it to null.

Editor components that set a custom toolbar are now expected to unset
it in their destroyed() lifecycle handler.